Get data for repository and API

+++ app/src/main/java/com/petamind/example/jetweathercleanarchitecture/data/model/WeatherItem.kt
@@ -5,7 +5,7 @@ data class WeatherItem(
     val dt: Int,
     val dt_txt: String,
     val main: Main,
-    val pop: Int,
+    val pop: Float,
     val rain: Rain,
     val sys: Sys,
     val visibility: Int,
